+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @syntax/storybook
 
+## 0.7.0
+
+### Minor Changes
+
+- a7863fc: Introduce Box component
+
+### Patch Changes
+
+- Updated dependencies [a7863fc]
+  - @cambly/syntax-core@1.19.0
+
+## 0.6.8
+
+### Patch Changes
+
+- Updated dependencies [969e0cc]
+  - @cambly/syntax-core@1.18.0
+
 ## 0.6.7
 
 ### Patch Changes